Try to improve AppCard highlight animation code
No performance gains whatsoever. Seems like as soon as an AnimationController is "started", the CPU usage skyrockets. Even the number of AnimationControllers *created* (not started) doesn't seem to matter. Luckily, only one of all those controllers (one is created per each AppCard) "runs" at a single time (when the AppCard is focused), maybe if that wasn't the case, things would be a lot worse.
Using a single AnimationController for all the AppCards made no difference.